moved blackBox function to its own script to modularize server-side code
